I don't know any more
Thread 1 "SpyPlayer" received signal SIGSEGV, Segmentation fault.
0x00000000005bae90 in AttachAudioStreamProcessor ()

+++ controls.odin
@@ -814,7 +814,7 @@ AudioVisualizerControl :: struct {
 	name:                string,
 	enabled:             bool,
 	positionRec:         raylib.Rectangle,
-	bars:                []int,
+	bars:                []f32,
 	barColors:           []raylib.Color,
 	tint:                raylib.Color,
 	texture:             raylib.Texture2D,
